REFACTOR, new class for jsdom
